fix(desktop): 修复无法记录MSLFrp登陆状态的bug style(desktop): 优化数据绑定代码
