+++ MSLX.Desktop/ViewModels/ViewModelBase.cs
@@ -1,0 +1,11 @@
+using CommunityToolkit.Mvvm.ComponentModel;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MSLX.Desktop.ViewModels
+{
+    public abstract class ViewModelBase : ObservableObject
+    {
+    }
+}
